+++ plugins/lime-plugin-mesh-wide/src/meshWideApi.ts
@@ -1,9 +1,15 @@
 import { QueryKey } from "@tanstack/react-query";
 
+import { callToRemoteNode } from "plugins/lime-plugin-mesh-wide-upgrade/src/utils/api";
+import {
+    getFromSharedStateKeys,
+    publishAllFromSharedStateAsyncKey,
+} from "plugins/lime-plugin-mesh-wide/src/meshWideQueriesKeys";
 import {
     DataTypeMap,
     DataTypes,
     SharedStateReturnType,
+    completeDataTypeKeys,
 } from "plugins/lime-plugin-mesh-wide/src/meshWideTypes";
 
 import api from "utils/uhttpd.service";
@@ -19,3 +25,41 @@ export const doSharedStateApiCall = async <T extends DataTypes>(
     }
     return res.data;
 };
+
+/**
+ * Sync all data types from shared state from remote node
+ */
+
+interface ISyncWithIpProps {
+    ip: string;
+}
+
+export async function publishOnRemoteNode({ ip }: ISyncWithIpProps) {
+    return await callToRemoteNode({
+        ip,
+        apiCall: (customApi) =>
+            customApi
+                .call(...publishAllFromSharedStateAsyncKey, {})
+                .then(() => true),
+    });
+}
+
+export async function syncDataType({
+    dataType,
+    ip,
+}: {
+    dataType: DataTypes;
+    ip: string;
+}) {
+    const queryKey = getFromSharedStateKeys.syncFromSharedStateAsync(dataType, [
+        ip,
+    ]);
+    return doSharedStateApiCall<typeof dataType>(queryKey);
+}
+
+export async function syncAllDataTypes({ ip }: { ip: string }) {
+    const promises = (Object.keys(completeDataTypeKeys) as DataTypes[]).map(
+        (dataType) => syncDataType({ dataType, ip })
+    );
+    return await Promise.all(promises);
+}
